requirements.txt
flair==0.10
nltk==3.6.2
numpy==1.21.6
openai==0.16.0
pandas==1.3.5
protobuf==3.20
python-Levenshtein==0.12.2
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.0
tqdm==4.62.3
transformers==4.21.2